README installation update
Updated to make installation easier and more accessible.

Converted the python script into an executable. 

How to use:

Download Stack folder
Open the dist folder
Run the Stack.exe
You should be prompted to enter the Directory that needs to be sorted.
Files should be organized based on the file extension.

Thats it!